feat: preparing accessing employee table
