+++ MyEMS/src/UIs/logInUI.java
@@ -1,6 +1,6 @@
 package UIs;
 import DB_Usage.User;
-import DB_Usage.emsDB;
+import DB_Usage.userInit;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionListener;
@@ -107,7 +107,7 @@ public class logInUI extends mainFrame implements ActionListener {
 
             String email = emailField.getText();
             String password = String.valueOf(passwordField.getPassword());
-            User user = emsDB.validateLogin(email, password);
+            User user = userInit.validateLogin(email, password);
 
             if (user != null) {
                 logInUI.this.dispose();
